認証関連処理をUserControllerからSignInControllerに切り分けました
