+++ app/views/users/sign_in.scala.html
@@ -13,7 +13,7 @@
 		<h1 class="text-center mt-5 fw-bold  py-3">ログイン</h1>
 		<div id="flashSuccessMessage" class="text-center text-success fw-bold my-3">
 			<span>@flash.get("success").getOrElse("")</span></div>
-		@form(routes.UserController.signIn()) {
+		@form(routes.SignInController.signIn()) {
 		@CSRF.formField
 		@signInForm.error("userNotFound") match {
 		case None =>{}
